remove: cleanup deprecated code and use testBoolQueryBuilderWrapper struct for testing
